Move BuildBucketAggregationQuery and BuildPartitionQuery to Helpers
